refactor: move logger JsonFormatter into utils directory
